requirements.txt
pandas>=2.0.0
chromadb>=0.4.0
langchain>=0.0.300
streamlit>=1.26.0
typer>=0.9.0
llama-cpp-python>=0.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
